+++ gui/src/components/widgets/__init__.py
@@ -1,3 +1,4 @@
+from gui.src.components.widgets.gallery_filter_sort_bar import GalleryFilterSortBar
 from gui.src.components.widgets.recent_directories_picker import RecentDirectoriesPicker
 from gui.src.components.widgets.resource_simulator_dashboard import (
     MetricCard,
@@ -10,6 +11,7 @@ from gui.src.components.widgets.toast_widget import ToastWidget
 from gui.src.components.widgets.toggle_switch import ToggleSwitch
 
 __all__ = [
+    "GalleryFilterSortBar",
     "MetricCard",
     "RecentDirectoriesPicker",
     "ResourceSimulatorDashboard",
@@ -19,3 +21,4 @@ __all__ = [
     "ToastWidget",
     "ToggleSwitch",
 ]
+